# Before 
```
julia> @benchmark survival($t.ultimate,50,80)
BenchmarkTools.Trial: 10000 samples with 991 evaluations.
 Range (min … max):  40.448 ns … 420.240 ns  ┊ GC (min … max): 0.00% … 85.84%
 Time  (median):     43.769 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   46.705 ns ±  26.898 ns  ┊ GC (mean ± σ):  4.40% ±  6.76%

       ▃██▆▃                                                    
  ▂▂▂▂▆██████▅▅▅▄▃▃▃▃▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  40.4 ns         Histogram: frequency by time           65 ns <

 Memory estimate: 304 bytes, allocs estimate: 1.
```
# After
```
 julia> @benchmark survival($t.ultimate,50,80)
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  21.648 ns … 38.824 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     21.858 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   22.063 ns ±  0.745 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▃███▅▂▁      ▁▂▂▃▁▁         ▂▁▂  ▁                         ▂
  ▇███████▆▇▅▅▆▄██████▇██▇█▆▇▇▆██████▇▇▃▆▅▅▅▃▄▃▄▁▃▅▃▃▄▁▁▄▄▃▄▅ █
  21.6 ns      Histogram: log(frequency) by time      25.1 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
 ```